refactor and decoupling of edition and export

+++ app/src/main/java/com/videonasocialmedia/videona/model/entities/editor/media/Audio.java
@@ -22,6 +22,7 @@ import java.util.ArrayList;
 /**
  * An audio media item that represents a file  (or part of a file) that can be used in project audio
  * track.
+ *
  * @see com.videonasocialmedia.videona.model.entities.editor.media.Media
  */
 public class Audio extends Media {
@@ -33,7 +34,7 @@ public class Audio extends Media {
      *
      * @see com.videonasocialmedia.videona.model.entities.editor.media.Media
      */
-    public Audio(String identifier, String iconPath, String mediaPath, long fileStartTime, long duration, ArrayList<User> authors, License license) {
+    public Audio(String identifier, String iconPath, String mediaPath, int fileStartTime, int duration, ArrayList<User> authors, License license) {
         super(identifier, iconPath, mediaPath, fileStartTime, duration, authors, license);
     }
 
@@ -42,7 +43,7 @@ public class Audio extends Media {
      *
      * @see com.videonasocialmedia.videona.model.entities.editor.media.Media
      */
-    public Audio(String identifier, String iconPath, String selectedIconPath, String title, String mediaPath, long fileStartTime, long duration, Transition opening, Transition ending, MediaMetadata metadata, ArrayList<User> authors, License license) {
+    public Audio(String identifier, String iconPath, String selectedIconPath, String title, String mediaPath, int fileStartTime, int duration, Transition opening, Transition ending, MediaMetadata metadata, ArrayList<User> authors, License license) {
         super(identifier, iconPath, selectedIconPath, title, mediaPath, fileStartTime, duration, opening, ending, metadata, authors, license);
     }
 }
